+++ app/Http/Controllers/API/EducationController.php
@@ -2,10 +2,11 @@
 
 namespace App\Http\Controllers\API;
 
-use App\Http\Controllers\Controller;
-
 use App\Models\Education;
+
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use Mews\Purifier\Facades\Purifier;
 
 class EducationController extends Controller
 {
@@ -27,7 +28,10 @@ class EducationController extends Controller
             'institution'=>'required|string|max:50',
             'start_date'=>'required|date',
             'end_date'=>'required|date',
+            'user_id'=>'required|exists:users,id'
         ]);
+        $data['degree'] = Purifier::clean($data['degree'], 'default');
+        $data['institution'] = Purifier::clean($data['institution'], 'default');
 
         $education=Education::create($data);
         return response()->json([
@@ -53,11 +57,14 @@ class EducationController extends Controller
     public function update(Request $request, Education $education)
     {
         $data=$request->validate([
-            'degree'=>'required|string|max:50',
-            'institution'=>'required|string|max:50',
-            'start_date'=>'required|date',
-            'end_date'=>'required|date',
+            'degree'=>'sometimes|required|string|max:50',
+            'institution'=>'sometimes|required|string|max:50',
+            'start_date'=>'sometimes|required|date',
+            'end_date'=>'sometimes|required|date',
+            'user_id'=>'sometimes|required|exists:users,id'
         ]);
+        $data['degree'] = Purifier::clean($data['degree'], 'default');
+        $data['institution'] = Purifier::clean($data['institution'], 'default');
 
         $education->update($data);
         return response()->json([
@@ -70,6 +77,9 @@ class EducationController extends Controller
 
     public function destroy(Education $education)
     {
+        if(!$education){
+            return response()->json(['message' => 'not found'], 404);
+        }
         $education->delete();
         return response()->json([
             'message' => 'Education deleted successfully',
