+++ app/Http/Controllers/API/ExperienceController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use App\Models\Experience;
 use Illuminate\Http\Request;
+use Mews\Purifier\Facades\Purifier;
 
 class ExperienceController extends Controller
 {
@@ -32,7 +33,10 @@ class ExperienceController extends Controller
             'company'=>'required|string|max:50',
             'start_date'=>'required|date',
             'end_date'=>'required|date',
+            'user_id'=>'required|exists:users,id'
         ]);
+        $data['role'] = Purifier::clean($data['role'], 'default');
+        $data['company'] = Purifier::clean($data['company'], 'default');
 
         $experience=Experience::create($data);
 
@@ -60,11 +64,18 @@ class ExperienceController extends Controller
     public function update(Request $request, Experience $experience)
     {
         $data=$request->validate([
-            'role'=>'required|string|max:50',
-            'company'=>'required|string|max:50',
-            'start_date'=>'required|date',
-            'end_date'=>'required|date',
+            'role'=>'sometimes|required|string|max:50',
+            'company'=>'sometimes|required|string|max:50',
+            'start_date'=>'sometimes|required|date',
         ]);
+
+        if(isset($data['role'])){
+            $data['role'] = Purifier::clean($data['role'], 'default');
+        }
+
+        if(isset($data['company'])){
+            $data['company'] = Purifier::clean($data['company'], 'default');
+        }
 
         $experience->update($data);
         return response()->json([
@@ -79,6 +90,9 @@ class ExperienceController extends Controller
      */
     public function destroy(Experience $experience)
     {
+        if(!$experience){
+            return response()->json(['message' => 'not found'], 404);
+        }
         $experience->delete();
         return response()->json([
             'message' => 'Experience deleted successfully',
